Move the shared code to src/find_sources.rs
